linux

+++ src/BenchmarkDotNet/Portability/Cpu/CpuInfo.cs
@@ -6,9 +6,9 @@
         public int? PhysicalProcessorCount { get; }
         public int? PhysicalCoreCount { get; }
         public int? LogicalCoreCount { get; }
-        public int? CurrentClockSpeed { get; } //MHz
+        public double? CurrentClockSpeed { get; } //MHz
 
-        public CpuInfo(string processorName, int? physicalProcessorCount, int? physicalCoreCount, int? logicalCoreCount, int? currentClockSpeed)
+        public CpuInfo(string processorName, int? physicalProcessorCount, int? physicalCoreCount, int? logicalCoreCount, double? currentClockSpeed)
         {
             ProcessorName = processorName;
             PhysicalProcessorCount = physicalProcessorCount;
